Move the rest of the result-handling functionality to BaseFuture
